separate node loading from run list expansion, etc.
Node loading may be the first API call made in a chef-client run; if it
fails, we may be unable to communicate with the server entirely.

Operations after the node is loaded, such as run list expansion can fail
for a variety of reasons and we want to be able to report these failures
to the Chef server so the user can easily track them.